Update first-entry.md
Hi Mariah,

I added the information that I had originally put on the my entries page. In general, this page is now very full of information so I'm planning on going back and organizing it a little bit but I just want to see what these changes look like first.