Update rrna-analysis.md
Hey Mariah,

I'm about 90% confident that I already uploaded some of the information for 16S rRNA but it's not there. I might have forgotten to hit save so I started entering it again but I'm going to try to troubleshoot on github to see if I can find the information I might have put in- ignore this for now because it could be repeated information.